Better background player detection
previously we were losing track of a player started through
`xap player run -d`. Now we create a pid file that we can use
to give a status and kill it.
