+++ command/player/main.go
@@ -1,14 +1,14 @@
 package player
 
 import (
+	"bytes"
 	"errors"
 	"fmt"
+	"io/ioutil"
 	"os"
-	"os/exec"
+	"strconv"
 
 	"github.com/urfave/cli"
-
-	"github.com/robertgzr/xap/mp"
 )
 
 var Command = cli.Command{
@@ -27,74 +27,27 @@ var statusCommand = cli.Command{
 	Name:  "status",
 	Usage: "inspect the status of a detached mpv command",
 	Action: func(_ *cli.Context) error {
-		return errors.New("not implemented")
-	},
-}
-
-var defaultMpvFlags = []string{
-	"--no-video",
-	"--no-sub",
-	"--gapless-audio=yes",
-	"--volume=70.0",
-}
-var runCommand = cli.Command{
-	Name:        "run",
-	Usage:       "Start an instance of mpv.",
-	Description: `Everything after "--" is passed as arguments to mpv.`,
-	Flags: []cli.Flag{
-		&cli.BoolFlag{
-			Name:  "detached, d",
-			Usage: "run in the background",
-		},
-		&cli.BoolFlag{
-			Name:  "no-defaults",
-			Usage: "do not run with default flags",
-		},
-		&cli.BoolFlag{
-			Name:  "verbose, v",
-			Usage: "print the mpv command",
-		},
-	},
-	Action: func(ctx *cli.Context) error {
-		args := append(ctx.Args(),
-			fmt.Sprintf("--input-ipc-server=%s", ctx.GlobalString("socket")),
-			"--idle",
-		)
-		if !ctx.Bool("no-defaults") {
-			args = append(args, defaultMpvFlags...)
-		}
-
-		cmd := exec.Command("mpv", args...)
-		if ctx.Bool("verbose") {
-			fmt.Println("+", cmd.Args)
-		}
-
-		if ctx.Bool("detached") {
-			if err := cmd.Start(); err != nil {
-				return err
-			}
-			fmt.Fprintf(os.Stdout, "Started mpv... (pid: %d)\n", cmd.Process.Pid)
-			return nil
-		}
-
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		return cmd.Run()
-	},
-}
-
-var stopCommand = cli.Command{
-	Name:  "stop",
-	Usage: "Stops the mpv instance",
-	Action: func(ctx *cli.Context) error {
-		c, err := mp.Connect(ctx)
+		pidStr, err := ioutil.ReadFile(pidFile)
 		if err != nil {
 			return err
 		}
-		if err := c.Quit(); err != nil {
+		pid, err := strconv.Atoi(string(pidStr))
+		if err != nil {
 			return err
 		}
-		fmt.Fprintf(os.Stdout, "Stopped mpv...\n")
+		p, err := os.FindProcess(pid)
+		if err != nil {
+			return err
+		}
+		cmdline, err := ioutil.ReadFile(fmt.Sprintf("/proc/%d/cmdline", p.Pid))
+		if err != nil {
+			return err
+		}
+		args := bytes.Split(cmdline, []byte{0})
+		if string(args[0]) != "mpv" {
+			return errors.New("exited")
+		}
+		fmt.Fprintf(os.Stdout, "Runnning (pid: %d)\n", p.Pid)
 		return nil
 	},
 }
